requirements.txt
Flask>=3.0.0
pymongo>=4.6.0
Pillow>=10.1.0
numpy>=1.22.0
tensorflow==2.20.0
kagglehub>=0.2.0
jupyter>=1.0.0
ipykernel>=7.0.0
matplotlib>=3.8.0
python-dotenv>=1.0.0
werkzeug>=3.0.1
